make the Anthropic approval gate executor-bound
