+++ app/for-ai-companies/page.tsx
@@ -118,7 +118,7 @@ export default async function ForAiCompaniesPage() {
             For production: a <a href="https://github.com/emiliaprotocol/emilia-protocol/blob/main/examples/async-signoff.mjs" style={{ color: color.gold, textDecoration: 'none' }}>high-volume async signoff example</a> (the gate is selective &mdash; the agent loop never blocks on a human), and an <a href="https://github.com/xai-org/xai-cookbook/pull/42" style={{ color: color.gold, textDecoration: 'none' }}>open recipe PR</a> on xAI&rsquo;s cookbook.
           </p>
           <p style={{ fontSize: 14, color: color.t3, lineHeight: 1.7, maxWidth: 640, margin: '0 0 28px' }}>
-            For Grok specifically, the <a href="https://github.com/emiliaprotocol/emilia-protocol/blob/main/examples/grok_guard.py" style={{ color: color.gold, textDecoration: 'none' }}>hardened Python guard</a> does the offline cryptographic check itself &mdash; it verifies the device signature in-process against a <strong>pinned</strong> signer key, bound to the exact requested action and single-use, so a server merely saying &ldquo;approved&rdquo; is never enough. It ships with a <a href="https://github.com/emiliaprotocol/emilia-protocol/blob/main/examples/tests/test_grok_guard_redteam.py" style={{ color: color.gold, textDecoration: 'none' }}>red-team regression suite</a> that re-runs six attack vectors on every change.
+            The <a href="https://github.com/emiliaprotocol/emilia-protocol/blob/main/examples/executor_approval_gate.py" style={{ color: color.gold, textDecoration: 'none' }}>vendor-neutral Python executor gate</a> verifies the authorization receipt in-process against a <strong>pinned</strong> operator key, binds it to the exact requested action, applies a local freshness window, and enforces single-use, so a server merely saying &ldquo;approved&rdquo; is never enough. It ships with a <a href="https://github.com/emiliaprotocol/emilia-protocol/blob/main/examples/tests/test_executor_approval_gate_redteam.py" style={{ color: color.gold, textDecoration: 'none' }}>red-team regression suite</a> that runs 21 checks across seven attack classes plus genuine controls.
           </p>
           <div style={{ display: 'flex', gap: 12, flexWrap: 'wrap' }}>
             <Link href="/mcp" className="ep-cta" style={cta.primary}>The MCP server &rarr;</Link>
